Workstation/text_modifier/main.go
package main

import (
    "fmt"
    "os"
    "regexp"
    "strconv"
    "strings"
)

func main() {
    // Check if the correct number of arguments is provided
    if len(os.Args) != 3 {
        fmt.Println("Usage: go run . <input_file> <output_file>")
        return
    }

    inputFile := os.Args[1]
    outputFile := os.Args[2]

    // Read the content of the input file
    content, err := os.ReadFile(inputFile)
    if err != nil {
        fmt.Println("Error reading file:", err)
        return
    }

    // Modify the text
    modifiedText := modifyText(string(content))

    // Write the modified text to the output file
    err = os.WriteFile(outputFile, []byte(modifiedText), 0644)
    if err != nil {
        fmt.Println("Error writing file:", err)
    }
}

// modifyText applies all modifications to the input text
func modifyText(text string) string {
    // Replace hexadecimal numbers
    text = regexp.MustCompile(`(\b[0-9A-Fa-f]+)\s*\(hex\)`).ReplaceAllStringFunc(text, func(match string) string {
        hexStr := strings.Fields(match)[0]
        decimal, _ := strconv.ParseInt(hexStr, 16, 0)
        return fmt.Sprintf("%d", decimal)
    })

    // Replace binary numbers
    text = regexp.MustCompile(`(\b[01]+)\s*\(bin\)`).ReplaceAllStringFunc(text, func(match string) string {
        binStr := strings.Fields(match)[0]
        decimal, _ := strconv.ParseInt(binStr, 2, 0)
        return fmt.Sprintf("%d", decimal)
    })

    // Handle (up), (low), (cap) modifications
    text = regexp.MustCompile(`(\b\w+)\s*\((up|low|cap)\)`).ReplaceAllStringFunc(text, func(match string) string {
        parts := strings.Fields(match)
        word := parts[0]
        command := parts[1]

        switch command {
        case "up":
            return strings.ToUpper(word)
        case "low":
            return strings.ToLower(word)
        case "cap":
            return strings.ToTitle(word)
        }
        return match
    })

    // Adjust spacing around punctuation
    text = regexp.MustCompile(`\s*([.,!?;:])\s*`).ReplaceAllString(text, "$1 ")

    // Change "a" to "an" before vowels
    text = regexp.MustCompile(`\b(a)\s+([aeiouh])`).ReplaceAllString(text, "an $2")

    return text
}

Test 1

/*Sure, let's break down the simplified version of the program step by step:
Structure
The program consists of two main parts:
The main function: This is the entry point of the program. It's responsible for handling the command-line arguments, 
	reading the input file, calling the modifyText function to modify the text, 
		and writing the modified text to the output file.
The modifyText function: This function takes the input text as a parameter and applies all the necessary modifications to it. 
	It uses regular expressions to find and replace specific patterns in the text.
Implementation
Let's go through the implementation of the main function:
The program first checks if the correct number of command-line arguments (input file and output file) are provided. 
	If not, it prints a usage message and exits.
It then retrieves the input file and output file names from the command-line arguments.
Next, it reads the content of the input file using os.ReadFile and stores it as a string.
After that, it calls the modifyText function and passes the input text to it. 
	The modifyText function modifies the text and returns the modified version.
Finally, it writes the modified text to the output file using os.WriteFile.
Now, let's look at the implementation of the modifyText function:
Inside the modifyText function, there are several regular expressions that match specific patterns in the text:
(\b[0-9A-Fa-f]+)\s*$$hex$$: Matches hexadecimal numbers followed by (hex)
(\b+)\s*$$bin$$: Matches binary numbers followed by (bin)
(\b\w+)\s*$$(up|low|cap)$$: Matches words followed by (up), (low), or (cap)
\s*([.,!?;:])\s*: Matches punctuation marks with surrounding spaces
\b(a)\s+([aeiouh]): Matches "a" followed by a space and a vowel
The function uses regexp.MustCompile to create regular expression objects for each pattern.
It then uses the ReplaceAllStringFunc method to replace the matched patterns with modified versions:
For hexadecimal and binary numbers, it converts them to decimal using strconv.
	ParseInt and replaces them with the decimal values.
For case modifications, it uses strings.ToUpper, strings.
	ToLower, or strings.ToTitle to modify the case of the matched words.
For punctuation adjustment, it adds a space after the punctuation mark using strings.ReplaceAllString.
For article modification, it replaces "a" with "an" using strings.ReplaceAllString.
Finally, the modified text is returned from the modifyText function.
In summary, the program takes an input file, reads its content, modifies the text based on specific rules, 
	and writes the modified text to an output file. The modifications include converting hexadecimal 
		and binary numbers to decimal, changing the case of words, adjusting punctuation, 
			and replacing "a" with "an" before vowels. */